refactor(common): di 의존성 각 모듈에서 담당하도록 변경
